backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-CORS==3.0.10
Mutagen==1.45.1
pathlib==1.0.1
python-dotenv==1.0.0